Adding Image Component.
In the src map I added an "images" folder. All images that are going to be used for this project must be placed in that folder. This gives the project a better structure.

In the image component is an image imported from the images folder, that will be used as a source for <img>.