Adding BellSFX Component.
In the src map I added an "SFX" folder. All sound effects that are going to be used for this project must be placed in that folder. This gives the project a better structure.

In the BellSFX component is an deskbell.mp3 imported from the SFX folder, that will be used as a url for <Sound>.